refactor(gjc): consolidate lifecycle into ouroboros.gjc and de-DI activation
Restructure the GJC lifecycle without behavior change:

- ouroboros/gjc/ is now the single home for GJC domain logic (pure — no
  printing, no subprocess): paths.py (every canonical artifact location),
  mcp.py (registration ownership judgment + atomic owned removal),
  bridge.py (compatibility-bridge rendering/judgment, absorbed from
  cli/commands/gjc_bridge.py), guide.py (routing-guide rendering/judgment,
  moved out of runtime_instruction_artifacts.py).
- cli/gjc_setup.py is the one orchestration module: narration, GJC CLI
  invocation, and the snapshot/rollback transaction. It absorbs
  cli/commands/setup_gjc_activation.py, replaces the 14-callable injection
  with a GjcSetupHost dataclass binding the five setup-owned seams, and
  drops the locals()-probing rollback in favor of explicit state.
- cli/stdio_mcp_probe.py extracts the dependency-free stdio MCP health
  probe from cli/commands/codex.py so GJC endpoint validation no longer
  imports from the codex command module.
- setup.py keeps only host binding; uninstall/refresh import ownership
  judgment directly from ouroboros.gjc — no private cross-module imports.

Validation: ruff, mypy, module-size ratchet OK; pytest tests/unit passes
(remaining failures are machine-load-dependent perf tests in untouched
persistence/orchestrator modules, reproducing independent of this change).
